+++ plutarch-ledger-api/test/V1.hs
@@ -2,6 +2,7 @@ module V1 (tests) where
 
 import Plutarch.Integer (PInteger)
 import Plutarch.LedgerApi.V1 qualified as PLA
+import Plutarch.LedgerApi.Value as Value
 import Plutarch.Positive (PPositive)
 import PlutusLedgerApi.V1.Orphans ()
 import Test.Tasty (TestTree, adjustOption, testGroup)
@@ -46,6 +47,7 @@ tests =
     , adjustOption (fewerTests 16) $ checkLedgerProperties @PLA.PTxInfo
     , checkLedgerPropertiesValue
     , checkLedgerPropertiesAssocMap
+    , checkLedgerProperties @Value.PAssetClass
     , testGroup
         "PCountable"
         [ checkLedgerPropertiesPCountable @PInteger